refactor(bookmark): move TargetType enum to shared package; update imports across bookmark module
