+++ shared/src/main/java/pt/estga/shared/enums/TargetType.java
@@ -1,4 +1,4 @@
-package pt.estga.mark.enums;
+package pt.estga.shared.enums;
 
 public enum TargetType {
     MARK,
